Add revision-history mode
Add versions data
Refactor detail-card
Make transitions smoother
Change header background and add fallbacks
