+++ addon/styles/detail-card.css
@@ -14,6 +14,7 @@
 
 .detail-card {
   width: 100%;
+  min-height: 100px;
   background-color: white;
   border-radius: 10px;
   box-shadow: var(--boxel-box-shadow);
