+++ app/components/detail-card.js
@@ -1,0 +1,1 @@
+export { default } from '@cardstack/boxel/components/detail-card';
