+++ tests/dummy/public/media-registry/api/profiles.json
@@ -36,6 +36,7 @@
     "id": "francesco-midina",
     "type": "profile",
     "title": "Francesco Midina",
+    "description": "Producer",
     "imgURL": "media-registry/profiles/francesco-midina.jpg",
     "organization": "verifi",
     "fields": [
