+++ tests/dummy/public/media-registry/api/songs_by_pia_midina_bb_clarke_table_1.json
@@ -270,7 +270,7 @@
     "parental_advisory": "No",
     "copyright_notice": "℗ 2020 Bunny Records",
     "congress_id": "SR0000320716",
-    "recording_date": "2019-01-16",
+    "recording_date": "2020-01-16",
     "original_release_date": "2020-03-02",
     "mastering_engineer": "Joel Kaplan",
     "mixing_engineer": "Mariah Solis",
